Move Utils.java from tachyon.hadoop (originally in the core module, now supposed to be in the client module) to tachyon.underfs.hdfs (the underfs module), because its method Utils.addS3Credentials() is needed in tachyon.hadoop.AbstractTFS (the hadoop client API in client module) and also tachyon.underfs.hdfs.HdfsUnderFileSystem.java (in the underfs module)
