+++ underfs/hdfs/src/main/java/tachyon/underfs/hdfs/HdfsUnderFileSystem.java
@@ -4,9 +4,9 @@
  * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance with the License. You may obtain a
  * copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software distributed under the License
  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
  * or implied. See the License for the specific language governing permissions and limitations under
@@ -40,7 +40,7 @@ import com.google.common.base.Throwables;
 import tachyon.conf.TachyonConf;
 import tachyon.Constants;
 import tachyon.TachyonURI;
-import tachyon.hadoop.Utils;
+import tachyon.underfs.hdfs.Utils;
 import tachyon.underfs.UnderFileSystem;
 import tachyon.util.NetworkUtils;
 
@@ -88,7 +88,7 @@ public class HdfsUnderFileSystem extends UnderFileSystem {
    * compatible implementations can override this method to add implementation specific
    * configuration necessary for obtaining a usable {@linkplain FileSystem} instance.
    * </p>
-   * 
+   *
    * @param path File system path
    * @param config Hadoop Configuration
    */
